make LengthError private to decl
