+++ src/data/mod.rs
@@ -3,8 +3,6 @@ use std::convert::TryFrom;
 use std::fmt::{self, Debug, Display, Formatter};
 use std::hash::Hash;
 
-use crate::arch::SIZE_T;
-
 pub mod error;
 pub mod lex;
 pub mod types;
@@ -181,32 +179,7 @@ impl Qualifiers {
     };
 }
 
-pub enum LengthError {
-    Unbounded,
-    Dynamic,
-    NonIntegral,
-    Negative,
-}
-
 impl Expr {
-    pub fn const_int(self) -> error::CompileResult<SIZE_T> {
-        use std::convert::TryInto;
-        if !self.ctype.is_integral() {
-            semantic_err!(LengthError::NonIntegral.into(), self.location,);
-        }
-        let literal = self.constexpr()?;
-        match literal.data.0 {
-            Literal::UnsignedInt(u) => Ok(u),
-            Literal::Int(x) => x.try_into().map_err(|_| {
-                CompileError::semantic(Locatable::new(
-                    LengthError::Negative.into(),
-                    literal.location,
-                ))
-            }),
-            Literal::Char(c) => Ok(u64::from(c)),
-            x => unreachable!("should have been caught already: {:?}", x),
-        }
-    }
     pub fn zero(location: Location) -> Expr {
         Expr {
             ctype: Type::Int(true),
@@ -214,25 +187,6 @@ impl Expr {
             expr: ExprType::Literal(Literal::Int(0)),
             lval: false,
             location,
-        }
-    }
-}
-
-impl From<LengthError> for String {
-    fn from(err: LengthError) -> String {
-        let s: &'static str = err.into();
-        s.to_string()
-    }
-}
-
-impl From<LengthError> for &'static str {
-    fn from(err: LengthError) -> &'static str {
-        use LengthError::*;
-        match err {
-            Unbounded => "Cannot take the length of unbounded array type",
-            Dynamic => "Length of variable-length array cannot be known at compile time",
-            NonIntegral => "The length of an array must be an integer",
-            Negative => "The length of an array must not be negative",
         }
     }
 }
